added cli test and conftest fixtures
